requirements.txt
ultralytics
torch
torchvision
opencv-python
pyyaml
numpy
